refactor(chain): extract identity lookup cache and TTL policy
